Moves existing integration tests to new package.
